requirements.txt
Django>=5.0.0
Pillow>=10.0.0
gunicorn
python-dotenv
whitenoise
django-tinymce